+++ Cargo.toml
@@ -41,6 +41,7 @@ rust-version = "1.76"
 ahash = "0.8.11"
 allocator-api2 = "0.2.16"
 anyhow = "1.0"
+arbitrary = "1.3.2"
 arrayref = "0.3.7"
 arrow = { version = "52.0.0" }
 arrow-arith = "52.0.0"
